Create standalone module for add-layers-url
...so it can be included into add-layers submodules
