+++ components/add-layers/arcgis/add-layers-arcgis.module.ts
@@ -5,10 +5,16 @@ import {TranslateModule} from '@ngx-translate/core';
 
 import {HsAddLayersArcGisComponent} from './add-layers-arcgis.component';
 import {HsAddLayersArcGisService} from './add-layers-arcgis.service';
+import {HsAddLayersCommonModule} from '../common/add-layers-common.module';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
-  imports: [CommonModule, FormsModule, TranslateModule],
+  imports: [
+    CommonModule,
+    FormsModule,
+    TranslateModule,
+    HsAddLayersCommonModule,
+  ],
   exports: [HsAddLayersArcGisComponent],
   declarations: [HsAddLayersArcGisComponent],
   providers: [HsAddLayersArcGisService],
