+++ components/add-layers/common/add-layers-url.directive.html
@@ -5,9 +5,9 @@
 
         <div class="input-group-append">
 
-            <button type="button" class="btn btn-secondary" [hidden]="!field.$valid" (click)="connect()"><i
+            <button type="button" class="btn btn-secondary" [hidden]="!url.valid" (click)="connect()"><i
                     class="icon-link"></i></button>
-            <button type="button" class="btn btn-danger" [hidden]="!field.$error.url"><i
+            <button type="button" class="btn btn-danger" [hidden]="!url.error.url"><i
                     class="icon-alertalt"></i></button>
             <hs-history-list [clicked]="historySelected" [what]="type"></hs-history-list>
         </div>
