+++ components/add-layers/vector/add-layers-vector.module.ts
@@ -3,13 +3,19 @@ import {CommonModule} from '@angular/common';
 import {FormsModule} from '@angular/forms';
 import {TranslateModule} from '@ngx-translate/core';
 
+import {HsAddLayersCommonModule} from '../common/add-layers-common.module';
 import {HsAddLayersVectorComponent} from './add-layers-vector.component';
 import {HsAddLayersVectorService} from './add-layers-vector.service';
 import {HsVectorUrlParserService} from './add-layers-vector-url-parser.service';
 
 @NgModule({
   schemas: [CUSTOM_ELEMENTS_SCHEMA],
-  imports: [CommonModule, FormsModule, TranslateModule],
+  imports: [
+    CommonModule,
+    FormsModule,
+    TranslateModule,
+    HsAddLayersCommonModule,
+  ],
   exports: [HsAddLayersVectorComponent],
   declarations: [HsAddLayersVectorComponent],
   providers: [HsAddLayersVectorService, HsVectorUrlParserService],
