+++ components/add-layers/wms/add-wms-layer.directive.html
@@ -1,5 +1,5 @@
 <form class="form-horizontal" name="frmConnect" role="form">
-    <hs-add-layers-url [type]="Wms" [url]="url" [connect]="connect" [field]="frmConnect.url"></hs-add-layers-url>
+    <hs-add-layers-url [type]="'Wms'" [url]="url" [connect]="connect"></hs-add-layers-url>
 
     <div [hidden]="!showDetails">
         <div class="form-group">
